Move ContactInfo to abook
