+++ kifi-backend/modules/abook/app/com/keepit/abook/ContactInfoRepo.scala
@@ -1,13 +1,17 @@
-package com.keepit.model
+package com.keepit.abook
 
 import com.google.inject.{Inject, Singleton, ImplementedBy}
-import com.keepit.common.db.slick.DBSession.RSession
-import com.keepit.common.db.slick._
-import com.keepit.common.db.{State, Id}
-import com.keepit.common.time.Clock
 import scala.slick.util.CloseableIterator
-
-// TODO: move this outside of shoebox
+import com.keepit.common.db.slick.Repo
+import com.keepit.model.ContactInfo
+import com.keepit.common.db.Id
+import com.keepit.model.User
+import com.keepit.common.db.slick.DBSession.RSession
+import com.keepit.common.db.slick.DataBaseComponent
+import com.keepit.common.time.Clock
+import com.keepit.common.db.slick.DbRepo
+import com.keepit.common.db.slick.FortyTwoTypeMappers
+import com.keepit.common.db.slick.DBSession
 
 @ImplementedBy(classOf[ContactInfoRepoImpl])
 trait ContactInfoRepo extends Repo[ContactInfo] {
@@ -35,4 +39,5 @@ class ContactInfoRepoImpl @Inject() (val db: DataBaseComponent, val clock: Clock
   def getByUserIdIter(userId: Id[User], maxRows:Int)(implicit session: RSession): CloseableIterator[ContactInfo] =
     (for(f <- table if f.userId === userId) yield f).elementsTo(maxRows)
 }
+
